Another try task

+++ lib/tasks/update_feed.rake
@@ -11,21 +11,25 @@ task :update_feed do
 end
 
 def get_news
-    base = "http://reshet.tv/news/channel2-news/daily-edition"
-    url = base
-    page = Nokogiri::HTML(open(url))
-    lastnews = page.to_s.scan(/http[^\"]*newsitem-[0-9]*/).uniq.sort.last
+    unless File.exist?('/tmp/last_news.mp4')
+        base = "http://reshet.tv/news/channel2-news/daily-edition"
+        url = base
+        page = Nokogiri::HTML(open(url))
+        lastnews = page.to_s.scan(/http[^\"]*newsitem-[0-9]*/).uniq.sort.last
 
-    url = URI.parse(lastnews.gsub('\\', ''))
-    page = Nokogiri::HTML(open(url))
-    lastvid = page.to_s.scan(/http[^\"]*mahadura[^\"]*mp4/).uniq.first.gsub('\\', '')
+        url = URI.parse(lastnews.gsub('\\', ''))
+        page = Nokogiri::HTML(open(url))
+        lastvid = page.to_s.scan(/http[^\"]*mahadura[^\"]*mp4/).uniq.first.gsub('\\', '')
 
-    open("/tmp/last_news.mp4", 'wb') do |file|
-        file << open(lastvid).read
+        open("/tmp/last_news.mp4", 'wb') do |file|
+            file << open(lastvid).read
+        end
     end
 
-    movie = FFMPEG::Movie.new("/tmp/last_news.mp4")
-    movie.transcode('/tmp/news_audio.aac', %w(-vn -acodec copy /tmp/news_audio.aac))
+    unless File.exist?('/tmp/news_audio.aac')
+        movie = FFMPEG::Movie.new("/tmp/last_news.mp4")
+        movie.transcode('/tmp/news_audio.aac', %w(-vn -acodec copy /tmp/news_audio.aac))
+    end
 
     s3 = Aws::S3::Resource.new
     obj = s3.bucket(ENV['BUCKET']).object('last_news')
